nvme: output also libnvme version
Print the used libnvme version which the nvme-cli fronent is using.

In order to be backward compatible don't expect that the newly
introduce function in libnvme to be 1available.

This is done by providing a weak symbol which is overwritten when
statically build or when libnvme is is used as shared libray the
dlsym() call checks if the function is available.

  $ nvme --version
  nvme version 2.0 (git 2.0-2-g037372a+)
  libnvme version 1.0 (git 1.0-2-gfb628fe)

Signed-off-by: Daniel Wagner <dwagner@suse.de>