build: Clarify current main project license
The original desired license for the project was GPL-2.0-or-later
but during the development a few files (e.g. fabrics.c and nvme.h)
got added which are licensed as GPL-2.0-only, which renders the
whole project to GPL-2.0-only.

Update the meson license field and the nvme.spec license field.
While at it mentions this in the contributions notes as well.

Signed-off-by: Daniel Wagner <dwagner@suse.de>