tests: Port to nose2 and unittest
The test are still using nose which has been superseeded by
nose2. Update the tests to use the nose2 and unittest.

To minimize errors in most of the porting is done using sed, e.g
's/TestNVMeIO.__init__(self)/super().setUp()/g'.

While at it also update the build system and move the test files to
the build dir. This avoids in tree exectution of the Python files
which leaves the __pycache__ dir in the source tree.

This also fixes the problem, that the config file is not found when
'meson test -C .build' is used. We have the prober dir structure for
Python and nose2 to find all resources in the build root. Furthermore,
when running the tests the output is captured in the nvmetests
dir. This dir is also created under the build root dir.

Signed-off-by: Daniel Wagner <dwagner@suse.de>