fabrics: Avoid out of bounds string chomping
First move nvme_strip_chomp() to fabrics.c as it is the only
user. Also rename it to strchomp as there is nothing nvme specific.

While at it, fix the implementation of the function to handle
all the corner cases by using strnlen() to figure out the
max length of the string. After we terminate it properly,
we can start chopping off any trailing whitespace.

Also make sure that all call sides of strchomp to use the correct max
lenght of the string to avoid out of bounds access.

Also we don't need space_strip_len() to tell use what the max size of
the string is. We know it at compile time.

Signed-off-by: Daniel Wagner <dwagner@suse.de>