mi: track controller lifetime under endpoint and provide endpoint scan
This change introduces a list of controllers under each MI endpoint, and
ties the controller's lifetime to the endpoints - when an endpoint is
free()ed, so are its controllers.

We also provide a new function to populate the controllers list:

   nvme_mi_scan_ep()

- which scans the endpoint for its owned controllers. For retrieving the
scan results, we now have a new macro:

   nvme_mi_for_each_ctrl(ep, ctrl) {
     ...
   }

Add a couple of new tests: one to check the lifetime semantics, and
another to test the scan using a fake Read MI Data (Controller List)
response.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>