types: fix typo on nvme_mi_ccs definitions in public API
The NVMe-MI spec defines the Composite Controller Status definitions
using the abbreviation CCS. However, the libnvme code (since initial
commit) has used CSS, not CCS (except for CCS_RDY...).

This change fixes the definitions to use the _CCS_ names, and adds a set
of #defines to allow the old names to be used, including for the enum
itself.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>