doc: Add a kdoc test for API docs
Even when we're not building docs, it would be handy to have an easy
check that the kernel-doc comments do not produce warnings or errors.

This change adds a test() defintion to docs/, which calls a simple
script, kernel-doc-check. This script invokes kernel-doc, looking for
error and/or warning output.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>