build: Remove auto review bot
Unfortunately, I was not able to figure out how to set the correct
permissions for necojackarc/auto-request-review to run.

The runner logs:

  Error: HttpError: Resource not accessible by integration

The documentation says that a PAT would work, but I really don't feel
comfortable to create such a powerful token.

So let's drop it and make the review assignment per manual
for the time being.

Links: https://github.com/necojackarc/auto-request-review#optional-github-personal-access-token

Signed-off-by: Daniel Wagner <dwagner@suse.de>