plugins/solidigm: fix return value on format parse failure
If validate_output_format() fails, there are a couple of paths in the
solidgm code that will return a (closed) file descriptor, rather than an
error status. This looks to be a couple of copy-and-paste errors.

This change fixes the return value for these failure path, ensuring that
we'll actually have an error return.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>