Change nvme_dev from global to static
We're currently exposing the global struct nvme_dev, to allow the old
users access to the name and statbuf data. Now that users have their
local reference (along with the fd) through parse_and_open instead, we
can drop the public definition, and declare it static to nvme.c.

We still keep the static declaration; this will be removed in an
upcoming change.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>