Introduce nvme_cli_ wrappers, wrap identify and identify_ctrl
This change introduces the first of the libnvme wrapper functions. These
have a nvme_cli_ prefix, and wrap the correspondingly-named libnvme
function.

This allows us to call the correct libnvme transport-specific function,
depending on whether this is a direct or MI device.

This requires an update to a more recent libnvme, for the mi version of
the identify functions.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>