fabrics: Remove dhchap-ctrl-secret from discover/connect-all
The dhchap-ctrl-secret is per controller but discover and connect-all
operate on several controllers.

As this get's very fast complicated how to express this on the command
line, just drop the it. It was not implemented anyway and not listed
in the documentation.

Use the JSON config file in this case.

Signed-off-by: Daniel Wagner <dwagner@suse.de>