mi: Introduce quirks and endpoint probing
This change introduces a little framework for adding device-specific
quirk flags. This allows the MI core to perform any queries on the
device, and set any quirk flags on specific devices.

This gives us a new (mi-internal) function:

    bool nvme_mi_ep_has_quirk(struct nvme_mi_ep *tp, unsigned long quirk)

which returns true if `quirk` is enabled on this endpoint.

Callers can disable probing through a new external API:

    nvme_mi_set_probe_enabled(nvme_root_t root, bool enabled);

Which will prevent sending any probe commands, and leave the quirks
field unset.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>