fabrics: Add ctrl to tracked list before doing discovering
TP8014 allows unique discovery controllers, this results in
should_connect() function no being able to figure out if controller
has been seen before. The problem is that we never connect (add_ctrl)
instead we do:

 do_discover
   connect_ctrls
      connect_ctrl
         do_discover
	 [...]

Do break this loop we need to make sure cargs_match_found() is able to
find a previous seen controller. Hence we add it to the tracked_ctrl
list in the above circle.

We are going to add it once more when we do a add_ctrl() but this not
a problem as add_ctrl() is not checking if the controller is already
in the list. The only place where the it matters is in
cargs_match_found(). So it the worst case we have a handful duplicates
in the list.

Signed-off-by: Daniel Wagner <dwagner@suse.de>