util/types: Add fallback for 128 bit types on 32bit platforms
gcc provides the __uint128_t type on only on 64bit platforms. A common
workaround is to use the gmp library in case longer arithmetic types
are needed.

As we don't do any math with these values (just reading and printing)
we can workaround by introducing a new type, nvme_uint128_t. We
only have two simple function operating on this type hence we just
have to make sure we use passing by value everwhere.

While at it we change the output format to hex based because this
avoids the overflow handling with the ten based system.

Signed-off-by: Daniel Wagner <dwagner@suse.de>