github workflows: Add an armhf cross build
This RFC change adds a github workflow for a cross-compile of libnvme on
armhf. We use a similar setup to the build-distro job, but install a
suitable cross compiler and provide a cross config to meson.

We currently setup with -Dpython=false, as the setup-python@v4 seems to
assume a native python build. We may be able to remove this later.

The suitability of this will depend on the resources available for
actions; doing some test runs on a private repo shows an execution time
of about 1 min 15 sec - a little less than the build-static job.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>