nvme, plugins: fix __u64 -> unsigned long long assumptions
We have a couple of instances which assume __u64 is an unsigned long,
which is not always the case.

For the printf() formats, we would ideally used PRId64, but on ll64
platforms, it looks this macro is broken: it expands to "ld", where
__u64 is an unsigned long long. So, do an explicit cast to unsigned long
long instead, and just use %lld.

With this change, we can compile without warnings on amd64, armhf and
powerpc64le.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>