fabrics: Add nvmf_get_discovery_wargs()
This function is similar to nvmf_get_discovery_log(), but takes
an extensible "args" parameter. The args parameter contains more
options than are available with nvmf_get_discovery_log(). For
example, it provides a way to set the LSP (Log page Specific field)
which is needed for advanced TP8010 features.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>