ioctl: Move nvme_get_log_page() to ioctl.h
Currently we use directly nvme_get_log() calls for many of the defined
helpers (e.g. nvme_get_log_persistent_event()). Though when the size
of the data is really large, nvme_get_log() is not able to transfer
the data (size exceeds what the kernel driver fit in a request).

In order to be able to use nvme_get_log_page() we have to move to
ioctl.h to avoid circular include dependency.

Signed-off-by: Daniel Wagner <dwagner@suse.de>