ioctl: Use nvme_get_log_page() in helpers
We use directly nvme_get_log() calls for many of the defined
helpers (e.g. nvme_get_log_persistent_event()). Though when the size
of the data is really large, nvme_get_log() is not able to transfer
the data (size exceeds what the kernel driver fit in a request).

Thus limit the transfer sizes in all helpers to 4k.

Signed-off-by: Daniel Wagner <dwagner@suse.de>