nvme: fw-download: Add retry mechanism
Since we're potentially sending a lot of (chunked) data in the
fw-download process, add a facility for a single chunk failure to be
retried, if the error status does not indicate the Do Not Resend status
flag.

The retry count is reset once a chunk is sucessfully transferred, up to
a maximum of three retries.

This allows for occasional transport errors to not abort the whole
transfer.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>