build: Add Azure Pipeline configuration
Verify if the change for nvme-cli is breaking blktests. We can't run
the blktests inside the containers because blktests wants write access
to sysfs and configfs which is forbidden in the provided github
containers. Thus we have to run our own fullfledge VM.

The instruction how to setup the build infrastructure is coming soon.

Signed-off-by: Daniel Wagner <dwagner@suse.de>