build: Disable Azure Pipeline
Unfortunatly, currently the build agent setup is not working
correctly. The issue is that when the build VM is spawned and the
agent is installed via the Extension API it will not show up as build
agent (though as deploy agent).

Until this is fixed, just disable this CI build.

Signed-off-by: Daniel Wagner <dwagner@suse.de>