build: Allow static linking
To allow static linking, we have to make the libnss dependency
optional. This means the hostname IP resolution feature is not
available in this configuration.

Also the libystemd dependency check doesn't take into account that we
want to link statically. Lukily we just need this for an example.

Currently, the OpenSSL library build will fail in the example
(dependency on libnss) but the library is available to link.

Signed-off-by: Daniel Wagner <dwagner@suse.de>